--- src/main/java/org/example/oenskelisten/Interface/IWishRepository.java
@@ -1,4 +1,0 @@
-package org.example.oenskelisten.Interface;
-
-public interface IWishRepository {
-}
