+++ src/main/resources/static/css/general.css
@@ -52,6 +52,25 @@ footer {
     height: 10vh;
 }
 
+/* GitHub icon and text container */
+.github-icon {
+    display: flex;
+    align-items: center;  /* Vertically align items */
+    justify-content: center;  /* Center items horizontally */
+    gap: 10px;  /* Space between the icon and the text */
+}
+
+/* Styling for the GitHub icon */
+.github-icon i {
+    font-size: 40px;  /* Adjust the size of the icon */
+}
+
+/* Styling for the text */
+.github-icon span {
+    font-size: 18px;  /* Adjust the font size for the text */
+    color: white;  /* Make the text white (or whatever color fits your design) */
+}
+
