+++ src/main/resources/templates/edit-wish.html
