+++ src/main/resources/templates/index.html
@@ -18,7 +18,7 @@
 
     <div>
     <label for="userLogin">Login</label>
-    <a th:href="@{/user-login}" id="userLogin">Login</a>
+    <a th:href="@{/session/user-login}" id="userLogin">Login</a>
     </div>
 </main>
