+++ src/main/resources/templates/user-form.html
@@ -19,20 +19,25 @@
 
         <div>
             <label for="email">Email:</label>
-            <input type="text" name="email" id="email" th:value="*{email}">
+            <input type="text" name="email" id="email" th:value="*{email}" required>
         </div>
 
         <div>
             <label for="password">Password:</label>
-            <input type="text" name="password" id="password" th:value="*{password}">
+            <input type="text" name="password" id="password" th:value="*{password}" required>
         </div>
 
         <div>
             <label for="name">Name:</label>
-            <input type="text" name="name" id="name" th:value="*{name}">
+            <input type="text" name="name" id="name" th:value="*{name}" required>
         </div>
 
-        <button type="submit">Opret</button>
+        <div>
+            <button type="submit">Opret</button>
+            <a th:href="@{'/'}">
+                <button type="button">Fortryd</button>
+            </a>
+        </div>
     </form>
 </main>
