+++ src/main/resources/templates/user-login.html
@@ -12,16 +12,15 @@ xmlns:th="http://www.thymeleaf.org">
 <main>
     <h1>Login</h1>
 
-    <form th:action="@{/login}" method="post">
+    <form th:action="@{/session/login}" method="post">
         <p th:if="${wrongCredentials}" class="error">Forkert email eller password</p>
-        <p th:if="${insertValue}" class="error">Udfyld email og password</p>
         <div>
             <label for="email">Email:</label>
-            <input id="email" name="email" type="text" autofocus>
+            <input id="email" name="email" type="text" required>
         </div>
 
         <div><label for="password">Password:</label>
-            <input id="password" name="password" type="password" autofocus>
+            <input id="password" name="password" type="password" required>
         </div>
         <div>
             <button type="submit">Login</button>
