+++ src/main/resources/templates/wish-list-item.html
