+++ src/main/resources/templates/wish-list-items.html
